requirements.txt
streamlit==1.41.1
tensorflow==2.15
pillow== 11.0.0
numpy==1.26.4
requests